Task関係のクラスにおいて、変数とメソッドの処理内容、名前変更。

+++ src/app/repository/TaskRepository.java
@@ -22,22 +22,25 @@ public class TaskRepository {
     // formatterを共通化
     private static final DateTimeFormatter DATETIME_FORMATTER =
         DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss");
+    // アプリ起動時のファイル有無
+    boolean existFile = false;
 
     // CSVファイルに追記
     public void saveTask(Task task){
         
         File file = new File(TASK_FILE_PATH);
-        // ファイル有無確認
-        boolean notExistFile = !file.exists();
+        // ファイル有無確認       
+        existFile = file.exists();
         
         try (BufferedWriter bw = new BufferedWriter(
             new OutputStreamWriter(
-                new FileOutputStream(TASK_FILE_PATH, true),
+                new FileOutputStream(TASK_FILE_PATH, true),     // true:追記モード
                 StandardCharsets.UTF_8))) {
                     
                     // CSVファイルが存在しない場合、またはファイルの中身が空である場合、ヘッダを１行目に挿入する。
-            if(notExistFile || file.length() == 0){
+            if(!existFile || file.length() == 0){
                 bw.write("task_id,taskName,category,createdAt,updatedAt,deletedAt\n");
+                existFile = true;
             }
             
             bw.write(task.toCsv() + "\n");
@@ -50,38 +53,34 @@ public class TaskRepository {
     
     // CSVファイル読み込み
     public List<Task> loadTasks(){
-        
+
+        // ファイル指定
         File file = new File(TASK_FILE_PATH);
         // ファイル有無確認
-        boolean notExistFile = !file.exists();
+        existFile = file.exists();
         
         // リスト（読み込みデータを入れる箱）準備
         List<Task> list = new ArrayList<>(); 
         
         // CSVファイルが存在しない場合、空のリストを戻しメソッドを終了させる
-        if(notExistFile){
+        if(!existFile){
             System.out.println("no Task data\n");
             return list;
         }
         try(BufferedReader br = new BufferedReader(
             new InputStreamReader(
                 new FileInputStream(TASK_FILE_PATH),
-                StandardCharsets.UTF_8))){
-            
+                StandardCharsets.UTF_8))){            
                     
             String line;
-
             // ファイルの終わりまで一行ずつ読む。
             while((line = br.readLine()) != null){
-
                 // 1行目（カラム名）を無視
                 if(line.startsWith("task_id")){
                     continue;
                 }
-
                 // ","区切りで文字列を分解
                 String[] data = line.split(",");
-
                 int taskId = Integer.parseInt(data[0]);
                 String task = data[1];
                 String category = data[2];
@@ -101,18 +100,29 @@ public class TaskRepository {
         return list;
     }
 
-    // CSVファイルを上書き更新
+    // CSVファイル上書き更新
     public void saveAll(List<Task> tList){
         
+        // ファイル指定
         File file = new File(TASK_FILE_PATH);
+        // ファイル有無確認
+        existFile = file.exists();
+            
+        // ファイルが無い場合、
+        if(!existFile){
+            System.out.println("CSVファイルが存在しません。");
+            return;
+        }
         
         try (BufferedWriter bw = new BufferedWriter(
             new OutputStreamWriter(
-                new FileOutputStream(TASK_FILE_PATH, false),
+                new FileOutputStream(TASK_FILE_PATH, false),    // false:上書きモード
                 StandardCharsets.UTF_8))) {
 
-            for(Task task : tList){
-                bw.write(task.toCsv() + "\n");
+            bw.write("task_id,taskName,category,createdAt,updatedAt,deletedAt\n");
+            
+            for(Task t : tList){
+                bw.write(t.toCsv() + "\n");
             }
             System.out.println("上書きしました");
         } catch (IOException e) {
